overhauled project structure
Overhauled the project structure.

Previously was a bunch of js files thrown onto the src folder without care. Now a pages, components, assets, etc. structure is in place.

App.js was rewritten to reflec the changes.
ONLY HomePage was refactored to use react-router-dom instead of navigateTo. This is because a navbar component will be implemented later on.